+++ docs/zh-CN/configurations.md
@@ -1,0 +1,3 @@
+# HyperMD 配置参数
+
+此文档暂时没有中文版。请参考[（由源代码自动生成的）英文文档](../configurations.md)
